compiles clean with -Wall
Ignore-this: e9a6e7c9fc6e58d2c884bddd8e8bfa75

Now everything compiles clean with -Wall -Werror.
Compiler found a couple of missed return statements. nice!

darcs-hash:20141001033919-c21fe-9cd355680661b01413ad4f8df816c7c4ba8615df

+++ src/gui/lua/base.cpp
@@ -544,7 +544,7 @@ namespace fire
 
                         //lua arrays start with 1
                         if(ik != 0) ik --;
-                        if(ik >= a->size()) a->resize(ik + 1);
+                        if(ik >= static_cast<int>(a->size())) a->resize(ik + 1);
 
                         (*a)[ik] = v;
                     }
